feat(integrations): split GitHub sanitized inputs check into two automations
Splits the combined "Sanitized Inputs & Code Scanning" GitHub check into
two separate IntegrationChecks so each renders as its own card in the
integrations UI:

- input_validation - "Input Validation" - scans for validation libraries
  in package.json / requirements.txt / pyproject.toml / composer.json
- code_scanning - "Code Scanning" - detects CodeQL default setup,
  custom CodeQL workflows, or third-party SARIF uploaders

Both checks remain mapped to the existing sanitizedInputs task template;
the runner aggregates pass/fail across checks per task, so auto-completion
behavior is preserved (task only completes when both pass for every repo).

Detection logic is unchanged - the JS/PY/PHP package catalogs and the
CodeQL detection flow (default-setup API + workflow file fallback) move
verbatim into the new files. Shared catalogs are extracted to
validation-libraries.ts; CodeQL detection helpers are extracted to
code-scanning-detector.ts to keep each file under the 300-line limit.
