+++ packages/integration-platform/src/manifests/github/checks/index.ts
@@ -4,6 +4,7 @@
  */
 
 export { branchProtectionCheck } from './branch-protection';
+export { codeScanningCheck } from './code-scanning';
 export { dependabotCheck } from './dependabot';
-export { sanitizedInputsCheck } from './sanitized-inputs';
+export { inputValidationCheck } from './input-validation';
 export { twoFactorAuthCheck } from './two-factor-auth';
