+++ packages/integration-platform/src/manifests/github/index.ts
@@ -7,8 +7,9 @@
 
 import type { IntegrationManifest } from '../../types';
 import { branchProtectionCheck } from './checks/branch-protection';
+import { codeScanningCheck } from './checks/code-scanning';
 import { dependabotCheck } from './checks/dependabot';
-import { sanitizedInputsCheck } from './checks/sanitized-inputs';
+import { inputValidationCheck } from './checks/input-validation';
 import { twoFactorAuthCheck } from './checks/two-factor-auth';
 
 export const manifest: IntegrationManifest = {
@@ -82,7 +83,13 @@ export const manifest: IntegrationManifest = {
   ],
 
   // Compliance checks that run daily and can auto-complete tasks
-  checks: [branchProtectionCheck, dependabotCheck, sanitizedInputsCheck, twoFactorAuthCheck],
+  checks: [
+    branchProtectionCheck,
+    codeScanningCheck,
+    dependabotCheck,
+    inputValidationCheck,
+    twoFactorAuthCheck,
+  ],
 
   isActive: true,
 };
